% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_geotherm.R
\name{addGeotherms}
\alias{addGeotherms}
\title{Add geotherms to an existing plot canvas}
\usage{
addGeotherms(p, geoth, curveVar = NULL, lwd = 1, colscale = NULL)
}
\arguments{
\item{p}{the plot canvas}

\item{geoth}{a tibble, containing the geotherm values. It requires at least
columns `Depth`, `Temperature` and the varying parameter (defined in `var`), corresponding to the color code.}

\item{curveVar}{the paramters that varies and defines the curve bundle. It can be automatically read from attributes}

\item{lwd}{the line width (constant)}

\item{colscale}{the color scale. Discrete. Make sure it has the same length as the number of cases.
If not supplied, it tries to give sensible defaults.}
}
\value{
a ggplot object containing the plot. It is returned invisibly so you need to plot it manually.
}
\description{
Add geotherms to an existing plot canvas
}
