% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readbackgrounds.R
\docType{data}
\name{faciesMeta}
\alias{faciesMeta}
\title{Metamorphic facies grid, within the crust}
\format{
An object of class \code{list} of length 6.
}
\usage{
faciesMeta
}
\description{
Metamorphic facies grid, within the crust
}
\keyword{datasets}
