% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readbackgrounds.R
\docType{data}
\name{faciesUltraMeta}
\alias{faciesUltraMeta}
\title{Metamorphic facies grid, ulra-metamorphism}
\format{
An object of class \code{list} of length 7.
}
\usage{
faciesUltraMeta
}
\description{
Metamorphic facies grid, ulra-metamorphism
}
\keyword{datasets}
