% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_geotherm.R
\name{plotCanvas}
\alias{plotCanvas}
\title{Prepare a canvas, with the appropriate background and axes}
\usage{
plotCanvas(bg)
}
\arguments{
\item{bg}{the background. A list containing the following:
* `image`, in a format that can be read by the system
* `Tmin`, `Tmax`, the temperature of the left and right border of the image
* `zmin`, `zmax`, the depth of the top and bottom border of the image, in km. 
   They are depth, so positive numbers, going down.
* `Tpos`, `zpos`, the position (and labels) of the ticks on each axis. `zpos` is in km. Positive,going down.}
}
\value{
(invisibly) a ggplot object with an empty canvas, including a background image and properly set-up axes.
The empty plot is returned invisibly, you need to plot it explicitly if you want it !
}
\description{
Prepare a canvas, with the appropriate background and axes
}
