% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readbackgrounds.R
\docType{data}
\name{React}
\alias{React}
\title{Reactions, pelitic system}
\format{
An object of class \code{list} of length 7.
}
\usage{
React
}
\description{
Reactions, pelitic system
}
\keyword{datasets}
