% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_geotherm.R
\name{steady_geotherm_bundle}
\alias{steady_geotherm_bundle}
\title{Calculate a series of geotherm, varying one of the input parameters
Parameters allowed to vary are z_moho, k, q_moho and HP}
\usage{
steady_geotherm_bundle(
  T_surf,
  rho,
  Cp,
  z_moho = 0,
  k = 0,
  q_moho = 0,
  HP = 0,
  varying,
  varies_from,
  varies_to,
  n_curves,
  N,
  UItoSI = NULL,
  model = heatModel
)
}
\arguments{
\item{T_surf}{surface temperature [°C] or [K]}

\item{rho}{Specific mass of the crust [km.m-3]}

\item{Cp}{Heat capacity of the crust [J.K-1.kg-1]}

\item{z_moho}{depth to the moho, in [m]}

\item{k}{Heat conductivity [W.m-1.K-1]}

\item{q_moho}{Heat flux at the moho [W.m-2]}

\item{HP}{Heat production of the rocks [W.m-3]}

\item{varying}{(string): the name of the parameter that varies. It changes from `varies_from`
to `varies_to`, generating a total of `n_curves` curves.
The value that is may have been supplied is ignored}

\item{varies_from}{minimum value of the varying parameter}

\item{varies_to}{maximum value of the varying parameter}

\item{n_curves}{total number of geotherms to compute}

\item{N}{Number of cells in the model}

\item{UItoSI}{the conversion factors from user-specified values to SI values.
This is a named vector with the names of the parameters, i.e. `z_moho`, `k`, `q_moho`, `rho`, `Cp` and `HP`.
If not supplied, the function will assume that they are all 1, i.e. values are in SI already.}

\item{model}{The modelling function}
}
\value{
a tibble in "long" format, with depth (in SI units), temperature and the parameters used for each curve
}
\description{
Calculate a series of geotherm, varying one of the input parameters
Parameters allowed to vary are z_moho, k, q_moho and HP
}
\note{
This function calculates in SI units, but the values can be specified in user-defined (UI) units
as long as the conversion factors UI to SI are supplied. Pay attention to mW, km, Ma, etc.
}
